+++ kubejs/server_scripts/tfc/recipes.js
@@ -332,6 +332,4 @@ const registerTFCRecipes = (event) => {
 		'#forge:tools/hammers',
 		'#forge:tools/saws'
 	]).id('tfc:shapeless/jar_lid')
-
-	event.replaceInput({ mod: 'tfc' }, 'minecraft:sugar', '#tfg:sugars')
 }