+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - Fixed being able to dupe materials by crafting tools and then recycling them (#1100) @Pyritie
 - Fixed issue where "melts into" tooltips and certain timers on Jade were missing (#1102) @Xikaro
 - Added Wine utensils recipes for GT (#1111) @JeanRdS
+- Fixed kaolin clay duping (#1099) @SpicyNoodle5
+- Fixed TFC sugar recipes so they can now use any sugar (#1099) @SpicyNoodle5
 
 ## [0.9.9] - 30.05.2025
 ### Changes